issue with spatial docs + major mine app document xref, also, a small ask from Denise
