+++ services/core-web/src/styles/components/PermitConditions.scss
@@ -8,7 +8,7 @@ div.condition-layer {
   .condition-content {
     &.editable:hover {
       cursor: pointer;
-      background-color: lightgrey;
+      background-color: #F4F0F0;
     }
 
     .view-item {
